requirements.txt
netaddr==0.7.18
PyYAML==3.10
yamllint==1.3.2
Jinja2==2.7.2
